+++ source/player.cpp
@@ -28,7 +28,7 @@ Player::Player() {
     _at_ceiling = false;
 }
 
-void Player::PlayerUpdate(const std::unordered_set<int> keys_pressed, const float delta_time, const Map map) {
+void Player::PlayerUpdate(const std::unordered_set<int> keys_pressed, const int delta_time, const Map map) {
     switch (_current_state) {
         case PlayerState::stand:
             _speedx = .0f;
@@ -126,11 +126,11 @@ void Player::PlayerUpdate(const std::unordered_set<int> keys_pressed, const floa
     return;
 }
 
-float Player::getPosX() const {
+int Player::getPosX() const {
     return _old_posx;
 }
 
-float Player::getPosY() const {
+int Player::getPosY() const {
     return _old_posy;
 }
